Move configs for history server to internal.config.
